Improve logging
Logging is now more controllable. Logging is opt-in on a class basis --
to use it, simply include the module into your class, and from there
you'll get a logger method which resolves to a Logger object. There are
a few root-level properties on the Quickebooks module you can set to
change how these Logger objects are created.