Filter now handles a 'filter set'
Some filters do not simply accept a scalar value, they actually contain sets of
values. For instance, one of the filters under Windows/Invoice is ContactIdSet.
The XML for this looks like this:

  <ContactIdSet>
    <Id>100</Id>
    <Id>200</Id>
    <Id>300</Id>
  </ContactIdSet>

whereas another filter might look like this:

  <StartPage>3</StartPage>

So, we need a way to generate this kind of data.